nouveau fichier : .gitignore 	modifié :         Makefile 	modifié :         src/listcli.c 	modifié :         src/listprod.c 	modifié :         src/saifac.c 	modifié :         src/saiprod.c 	modifié :         src/saiscli.c 	modifié :         user-vergifac/lance-vergifac.sh 	modifié :         user-vergifac/vergifac 	supprimé :        user-vergifac/vergifac64 	modifié :         util/gencbox.c 	modifié :         util/genctrl.c 	modifié :         util/genfen.h 	modifié :         util/genfq.c 	modifié :         util/genftablo.c

+++ util/genctrl.c
@@ -314,7 +314,7 @@ for (i = 0, j = ll  ; j; ++ptf)  {
       }
 }
 
-long arron2(double ff, long long int ll)
+long long int arron2(double ff, long long int ll)
 {
 double  arron;
 double dmont;
